Final Year Btech Project
The aim of the paper is to identify consumer behaviour, perception and attitude towards energy drinks (ED). This article uses a questionnaire method to gather data and the selection of the study sample was random and involved a group of 330 people of various age, gender, yearly income and cities classified on the basis of their tier.
The result of this report gives us the correlation between people belonging to different gender, income bracket, residential setting and their energy drink consumption behaviour patterns. Also, SWOT analysis and strategies for enhancing customer acceptance for energy drinks in India are discussed.